# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inner Circle Partners Portal API",
    "description": "API documentation for the Inner Circle Partners Portal - a comprehensive partner referral management system.\n\n## Authentication\n\nMost endpoints require authentication via Clerk. Include the session token in your requests.\n\n## Rate Limiting\n\nAPI requests are rate-limited to prevent abuse:\n- General endpoints: 100 requests per minute\n- Auth endpoints: 10 requests per minute\n- Webhook endpoints: 1000 requests per minute\n\n## Response Format\n\nAll responses follow a standard format:\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"meta\": { \"timestamp\": \"...\" }\n}\n```\n\nError responses:\n```json\n{\n  \"success\": false,\n  \"error\": {\n    \"code\": \"ERROR_CODE\",\n    \"message\": \"Human readable message\"\n  }\n}\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Inner Circle Support",
      "email": "support@innercircle.co"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://innercircle.co/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.innercircle.co",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/admin/bulk": {
      "post": {
        "tags": ["Admin"],
        "summary": "Bulk operations",
        "description": "Perform bulk actions on partners or payouts",
        "operationId": "adminBulkAction",
        "security": [{ "clerkAuth": [], "adminRole": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action", "ids"],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "approve_partners",
                      "suspend_partners",
                      "activate_partners",
                      "approve_payouts",
                      "reject_payouts"
                    ]
                  },
                  "ids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "minItems": 1
                  },
                  "tier": {
                    "type": "string",
                    "enum": ["STANDARD", "SILVER", "GOLD", "PLATINUM"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk operation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkOperationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/export": {
      "get": {
        "tags": ["Admin"],
        "summary": "Export data",
        "description": "Export partners, payouts, or referrals as CSV",
        "operationId": "adminExport",
        "security": [{ "clerkAuth": [], "adminRole": [] }],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["partners", "payouts", "referrals"]
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file",
            "content": {
              "text/csv": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token"
      },
      "adminRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Role",
        "description": "Requires admin role in Clerk user metadata"
      }
    },
    "schemas": {
...
      "BulkOperationResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "total": { "type": "integer" },
              "successful": { "type": "integer" },
              "failed": { "type": "integer" },
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}